* src/mp3.h, src/mp3.cpp: Fix mpeg_frame_duration to use the static   array of block sizes instead of hard-coded almost-matching values.   Make mpeg_frame_length return a floating point value, and in   ReadHeader calculate total duration using floating point. Fixes   test-mp3-media-properties.html * test/media/audio/drtlist.xml: Added   test-mp3-progressive-download.html. * test/media/audio/miguel-11.025kHz-16kbps.mp3,   test/media/audio/miguel-11.025kHz-16kbps-p.mp3,   test/media/audio/miguel-16kHz-56kbps.mp3,   test/media/audio/miguel-44.1kHz-56kbps.mp3,   test/media/audio/miguel-encode.sh,   test/media/audio/test-mp3-progressive-download.html: Added. * test/media/audio/test-mp3-media-properties.html: Improved to test all   miguel*.mp3 files. * test/harness/test-runner/Test.cs: RunLocationServerIfNeeded: add /:.   as xsp application, allows tests to access files in other   directories (test/js/js/*.js specifically).
svn path=/trunk/moon/; revision=115216

+++ src/mp3.h
@@ -120,6 +120,6 @@ public:
 };
 
 bool mpeg_parse_header (MpegFrameHeader *mpeg, const guint8 *buffer);
-guint32 mpeg_frame_length (MpegFrameHeader *mpeg, bool xing);
+double mpeg_frame_length (MpegFrameHeader *mpeg, bool xing);
 
 #endif
